Port audit_homepage method to rubocop and add tests
