+++ Library/Homebrew/rubocops.rb
@@ -2,3 +2,4 @@ require_relative "./rubocops/bottle_block_cop"
 require_relative "./rubocops/formula_desc_cop"
 require_relative "./rubocops/components_order_cop"
 require_relative "./rubocops/components_redundancy_cop"
+require_relative "./rubocops/homepage_cop"
